requirements.txt
SQLAlchemy==2.0.42
psycopg2-binary==2.9.10

